% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/householder.R
\name{householder}
\alias{householder}
\title{Householder reflection}
\usage{
householder(x)
}
\arguments{
\item{x}{numeric vector of length 2 with x[1] > 0}
}
\value{
2x2 matrix H with H %*% x = -||x|| e1
}
\description{
Householder reflection
}
