% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{matmul_col_perspective}
\alias{matmul_col_perspective}
\title{Column Perspective Matrix Multiplication}
\usage{
matmul_col_perspective(A, B)
}
\arguments{
\item{A}{A numeric matrix.}

\item{B}{A numeric matrix.}
}
\value{
A numeric matrix representing the product of A and B.
}
\description{
Computes the matrix product by viewing the operation as a series of
matrix-vector products with the columns of B.
}
