% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{matmul_linear_transform}
\alias{matmul_linear_transform}
\title{Linear Transformation Composition Matrix Multiplication}
\usage{
matmul_linear_transform(A, B)
}
\arguments{
\item{A}{A numeric matrix.}

\item{B}{A numeric matrix.}
}
\value{
A numeric matrix representing the product of A and B.
}
\description{
Computes the matrix product by treating A and B as linear transformations
and evaluating their sequential functional composition on the standard basis vectors.
}
