% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{matmul_outer_product}
\alias{matmul_outer_product}
\title{Outer Product Matrix Multiplication}
\usage{
matmul_outer_product(A, B)
}
\arguments{
\item{A}{A numeric matrix.}

\item{B}{A numeric matrix.}
}
\value{
A numeric matrix representing the product of A and B.
}
\description{
Computes the matrix product by summing up the rank-1 outer products
of the columns of A and the rows of B.
}
