% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{matmul_row_perspective}
\alias{matmul_row_perspective}
\title{Row Perspective Matrix Multiplication}
\usage{
matmul_row_perspective(A, B)
}
\arguments{
\item{A}{A numeric matrix.}

\item{B}{A numeric matrix.}
}
\value{
A numeric matrix representing the product of A and B.
}
\description{
Computes the matrix product by viewing the rows of the output matrix
as linear combinations of the rows of B.
}
